update to bootworm

+++ uninstall.sh
@@ -2,17 +2,38 @@
 # uninstall minitowerkit script 
 . /lib/lsb/init-functions
 
-moodlight_svc="minitower_moodlight.service"
-filelocation="/lib/systemd/system/${moodlight_svc}.service"
+log_action_msg "Uninstalling minitower driver..."
+sleep 3
+log_action_msg "Disable moodlight and oled services."
+sudo systemctl disable minitower_moodlight.service
+sudo systemctl disable minitower_oled.service
 
-log_action_msg "Uninstalling minitower moodlight Driver..."
-sleep 1
-log_action_msg "Remove dtoverlay configure from /boot/config.txt file"
-sudo sed -i '/dtparam=i2c.*/ s/^/#/' /boot/config.txt
-log_action_msg "Stop and disable ${moodlight_svc}"
-sudo systemctl disable ${moodlight_svc} 2&>/dev/null  
-sudo systemctl stop ${moodlight_svc}  2&>/dev/null
-log_action_msg "Remove Minitower kit Driver..."
-sudo rm -f  /lib/systemd/system/minitower_moodlight.service  2&>/dev/null 
-sudo rm -f /usr/bin/moodlight 2&>/dev/null 
+log_action_msg "Stop moodlight and oled services."
+sudo systemctl stop minitower_moodlight.service
+sudo systemctl stop minitower_oled.service
+
+log_action_msg "Remove moodlight and oled service files."
+if [[ -e /etc/systemd/system/minitower_moodlight.service ]];then
+   sudo rm -f /etc/systemd/system/minitower_moodlight.service
+fi 
+
+if [[ -e /etc/systemd/system/minitower_oled.service ]];then
+   sudo rm -f /etc/systemd/system/minitower_oled.service
+fi
+
+log_action_msg "Remove dtoverlay configure from /boot/firmware/config.txt file"
+sudo sh -c "sed -i '/dtparam=i2c.*/ s/^/#/' /boot/firmware/config.txt"
+
+log_action_msg "Remove /usr/bin/moodlight file."
+if [[ -e /usr/bin/moodlight ]]; then
+	sudo rm -f /usr/bin/moodlight 2&>/dev/null 
+fi
+
+log_action_msg "Remove minitower oled python script"
+if [[ -d /usr/local/minitower ]]; then
+	sudo rm -rf /usr/local/minitower/ 2&>/dev/null
+fi
+
+
 log_success_msg "Uninstall Mini tower kit Driver Successfully." 
+
